+++ findbugs-exclude-filter.xml
@@ -44,7 +44,7 @@
     </Match>
 
     <Match>
-        <Class name="jgnash.ui.report.FontRegistry"/>
+        <Class name="jgnash.report.ui.FontRegistry"/>
         <Method name="registerFontDirectories"/>
         <Bug pattern="DMI_HARDCODED_ABSOLUTE_FILENAME"/>
     </Match>
