+++ jgnash-fx/src/main/java/jgnash/uifx/report/PortfolioReportController.java
@@ -31,9 +31,9 @@ import jgnash.engine.AccountType;
 import jgnash.engine.CurrencyNode;
 import jgnash.engine.InvestmentPerformanceSummary;
 import jgnash.engine.SecurityNode;
-import jgnash.ui.report.jasper.AbstractReportTableModel;
-import jgnash.ui.report.jasper.ColumnHeaderStyle;
-import jgnash.ui.report.jasper.ColumnStyle;
+import jgnash.report.ui.jasper.AbstractReportTableModel;
+import jgnash.report.ui.jasper.ColumnHeaderStyle;
+import jgnash.report.ui.jasper.ColumnStyle;
 import jgnash.uifx.control.AccountComboBox;
 import jgnash.uifx.report.jasper.DynamicJasperReport;
 import jgnash.util.Nullable;
