+++ jgnash-fx/src/main/java/jgnash/uifx/report/jasper/AbstractCrosstabReport.java
@@ -28,9 +28,9 @@ import jgnash.engine.CurrencyNode;
 import jgnash.engine.Engine;
 import jgnash.engine.EngineFactory;
 import jgnash.time.DateUtils;
-import jgnash.ui.report.jasper.AbstractReportTableModel;
-import jgnash.ui.report.jasper.ColumnHeaderStyle;
-import jgnash.ui.report.jasper.ColumnStyle;
+import jgnash.report.ui.jasper.AbstractReportTableModel;
+import jgnash.report.ui.jasper.ColumnHeaderStyle;
+import jgnash.report.ui.jasper.ColumnStyle;
 import jgnash.uifx.control.DatePickerEx;
 import jgnash.resource.util.ResourceUtils;
 import net.sf.jasperreports.engine.JasperPrint;
