+++ jgnash-fx/src/main/java/jgnash/uifx/report/jasper/AbstractSumByTypeReport.java
@@ -26,10 +26,10 @@ import jgnash.engine.CurrencyNode;
 import jgnash.engine.Engine;
 import jgnash.engine.EngineFactory;
 import jgnash.time.DateUtils;
-import jgnash.ui.report.Row;
-import jgnash.ui.report.jasper.AbstractReportTableModel;
-import jgnash.ui.report.jasper.ColumnHeaderStyle;
-import jgnash.ui.report.jasper.ColumnStyle;
+import jgnash.report.ui.Row;
+import jgnash.report.ui.jasper.AbstractReportTableModel;
+import jgnash.report.ui.jasper.ColumnHeaderStyle;
+import jgnash.report.ui.jasper.ColumnStyle;
 import jgnash.uifx.control.DatePickerEx;
 import jgnash.resource.util.ResourceUtils;
 import net.sf.jasperreports.engine.JasperPrint;
