+++ jgnash-report-core/build.gradle
@@ -35,6 +35,6 @@ dependencies {
 
 jar {
     manifest {
-        attributes("Automatic-Module-Name": 'jgnash-report-core')
+        attributes("Automatic-Module-Name": 'jgnash.report')
     }
 }
