+++ jgnash-report-core/src/main/java/jgnash/report/ui/jasper/AWTFontUtilities.java
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package jgnash.ui.report.jasper;
+package jgnash.report.ui.jasper;
 
 import ar.com.fdvs.dj.domain.Style;
