+++ jgnash-report-core/src/main/java/jgnash/report/ui/jasper/ColumnStyle.java
@@ -15,32 +15,25 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package jgnash.ui.report;
-
-import jgnash.util.Nullable;
+package jgnash.report.ui.jasper;
 
 /**
- * Support interface to wrap a row of table data into one object
+ * Available report column styles
  *
  * @author Craig Cavanaugh
  */
-public abstract class Row<T> {
-
-    private final T object;
-
-    protected Row(@Nullable T object) {
-        this.object = object;
-    }
-
-    public T getValue() {
-        return object;
-    }
-
-    /**
-     * Returns the value given a column index
-     *
-     * @param columnIndex column index
-     * @return column value
-     */
-    public abstract Object getValueAt(final int columnIndex);
+public enum ColumnStyle {
+    AMOUNT_SUM,
+    BALANCE,
+    BALANCE_WITH_SUM,
+    BALANCE_WITH_SUM_AND_GLOBAL,
+    GROUP,
+    GROUP_NO_HEADER,
+    PERCENTAGE,
+    QUANTITY,
+    SHORT_AMOUNT,
+    SHORT_DATE,
+    STRING,
+    CROSSTAB_TOTAL,
+    TIMESTAMP
 }
