+++ jgnash-report-core/src/main/java/jgnash/report/ui/jasper/ReportFactory.java
@@ -15,14 +15,14 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package jgnash.ui.report.jasper;
+package jgnash.report.ui.jasper;
 
 import ar.com.fdvs.dj.domain.constants.Font;
 
 import java.text.NumberFormat;
 import java.util.prefs.Preferences;
 
-import jgnash.ui.report.FontRegistry;
+import jgnash.report.ui.FontRegistry;
 
 /**
  * Factory methods to help with report configuration and generation
