+++ jgnash-swing/src/main/java/jgnash/ui/report/compiled/AbstractSumByTypeReport.java
@@ -28,9 +28,9 @@ import jgnash.engine.Engine;
 import jgnash.engine.EngineFactory;
 import jgnash.time.DateUtils;
 import jgnash.ui.components.DatePanel;
-import jgnash.ui.report.jasper.AbstractReportTableModel;
-import jgnash.ui.report.jasper.ColumnHeaderStyle;
-import jgnash.ui.report.jasper.ColumnStyle;
+import jgnash.report.ui.jasper.AbstractReportTableModel;
+import jgnash.report.ui.jasper.ColumnHeaderStyle;
+import jgnash.report.ui.jasper.ColumnStyle;
 import jgnash.ui.report.jasper.DynamicJasperReport;
 import jgnash.ui.util.IconUtils;
 import jgnash.resource.util.ResourceUtils;
