+++ jgnash-swing/src/main/java/jgnash/ui/report/compiled/PortfolioReport.java
@@ -35,9 +35,9 @@ import jgnash.engine.CurrencyNode;
 import jgnash.engine.InvestmentPerformanceSummary;
 import jgnash.engine.SecurityNode;
 import jgnash.ui.components.AccountListComboBox;
-import jgnash.ui.report.jasper.AbstractReportTableModel;
-import jgnash.ui.report.jasper.ColumnHeaderStyle;
-import jgnash.ui.report.jasper.ColumnStyle;
+import jgnash.report.ui.jasper.AbstractReportTableModel;
+import jgnash.report.ui.jasper.ColumnHeaderStyle;
+import jgnash.report.ui.jasper.ColumnStyle;
 import jgnash.ui.report.jasper.DynamicJasperReport;
 import jgnash.ui.util.IconUtils;
